[Platinum V] Title: 비용, Time: 136 ms, Memory: 5132 KB -BaekjoonHub

+++ 백준/Platinum/2463. 비용/비용.cc
@@ -71,7 +71,7 @@ void get_result(){
             all -= n.value;
             continue;
         }
-        result += all*(int64_t)(a.getsize(n.from)*a.getsize(n.to));
+        result += all*(a.getsize(n.from)*a.getsize(n.to));   
         result %= 1000000000;
         a.marge(n.from, n.to);
         all -= n.value;
